Move all output util functions to boss.core.output
